Update lotj_channels.xml
The "Incoming/Ougoing" part of clan messaging was not going to the output window. It's convenient to know if a message is internal or external at a glance. I think your intention was for this to show up in the window, since there is a trigger for it. The biggest obstacle is the "$" in the regular expression which, I'm sure you already know, evaluates as "end of line." With a "$" at the beginning of the trigger string, it will never fire.

However, that alone is not a enough to make the [incoming/outgoing] line show up. I had to add some SetTriggerOption lines for clan2 as well, basically just mirroring what was done for say1/say2.